Moved helper looper into its own common pool
