+++ integration_test/src/python/integration_test/common/BUILD
@@ -6,6 +6,7 @@ pex_library(
     name = "heron-integration-common-py",
     srcs = glob(["**/*.py"]),
     deps = [
-        "//heronpy/api:heron-python-py"
+        "//heronpy/api:heron-python-py",
+        "//heronpy/dsl:heron-pythondsl-py"
     ],
 )
