+++ integration_test/src/python/integration_test/common/generator/__init__.py
@@ -1,0 +1,4 @@
+"""Integration test common generator"""
+__all__ = ['sleeparraylooper']
+
+from .sleeparraylooper import SleepArrayLooper
