+++ integration_test/src/python/integration_test/common/generator/sleeparraylooper.py
@@ -1,0 +1,37 @@
+# copyright 2016 twitter. all rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+'''SleepArrayLooper'''
+
+import collections
+import logging
+import itertools
+import time
+
+from heronpy.dsl.generator import Generator
+
+class SleepArrayLooper(Generator):
+  """A SleepArrayLooper loops the contents of the a user supplied array forever
+  """
+  def __init__(self, user_iterable):
+    super(SleepArrayLooper, self).__init__()
+    if not isinstance(user_iterable, collections.Iterable):
+      raise RuntimeError("SleepArrayLooper must be passed an iterable")
+    self._user_iterable = user_iterable
+
+  def setup(self, context):
+    self._curiter = itertools.cycle(self._user_iterable)
+
+  def get(self):
+    time.sleep(1)
+    return self._curiter.next()
